# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.5.10",
    "title": "META API",
    "description": "API文件"
  },
  "host": "six-northnode.onrender.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "jwt token"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/member/signup": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "會員註冊",
        "description": "會員註冊",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "required": "true",
            "description": "會員註冊用",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "example@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "password"
                },
                "nickName": {
                  "type": "string",
                  "example": "nickname"
                }
              },
              "required": [
                "email",
                "password",
                "nickName"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "註冊成功",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "token"
                    },
                    "createRes": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string",
                          "example": "example@gmail.com"
                        },
                        "profilePic": {
                          "type": "string",
                          "example": "/images/profilePic.jpeg"
                        },
                        "nickName": {
                          "type": "string",
                          "example": "nickname"
                        },
                        "_id": {
                          "type": "string",
                          "example": "mongoId"
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2023-04-26T15:47:12.219Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "2023-04-26T15:47:12.219Z"
                        },
                        "__v": {
                          "type": "number",
                          "example": 0
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "密碼強度",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "密碼長度至少8位、須包含數字與英文"
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "isOperational": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "402": {
            "description": "信箱、密碼、暱稱不可空白",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "信箱、密碼、暱稱不可空白"
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 402
                    },
                    "isOperational": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "406": {
            "description": "信箱重複",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "信箱重複"
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 406
                    },
                    "isOperational": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/member/signin": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "會員登入",
        "description": "會員登入",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "required": "true",
            "description": "會員登入用",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "example@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "password"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "登入成功",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "token"
                    },
                    "signinRes": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string",
                          "example": "example@gmail.com"
                        },
                        "password": {
                          "type": "string",
                          "example": "null"
                        },
                        "profilePic": {
                          "type": "string",
                          "example": "/images/profilePic.jpeg"
                        },
                        "nickName": {
                          "type": "string",
                          "example": "nickname"
                        },
                        "_id": {
                          "type": "string",
                          "example": "mongoId"
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2023-04-26T15:47:12.219Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "2023-04-26T15:47:12.219Z"
                        },
                        "__v": {
                          "type": "number",
                          "example": 0
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "email格式錯誤",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "信箱格式錯誤"
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "isOperational": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "402": {
            "description": "帳號或密碼欄位空白",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "帳號密碼必填"
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 402
                    },
                    "isOperational": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "密碼錯誤",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "密碼錯誤"
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "isOperational": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/member/checkEmail": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "驗證信箱重複",
        "description": "驗證信箱重複",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "required": "true",
            "description": "驗證信箱重複用",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "example@gmail.com"
                }
              },
              "required": [
                "email"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "email可以使用",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "example@gmail.com 可以使用"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "信箱格式錯誤",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "信箱格式錯誤"
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "isOperational": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "406": {
            "description": "信箱重複",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "信箱重複"
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 406
                    },
                    "isOperational": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/member/changePassword": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "修改會員密碼",
        "description": "修改會員密碼",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "required": "true",
            "description": "修改會員密碼用",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "password"
                },
                "confirmPassword": {
                  "type": "string",
                  "example": "password"
                }
              },
              "required": [
                "password",
                "confirmPassword"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "回傳物件",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "644cce67945042a407ed1c21"
                    },
                    "email": {
                      "type": "string",
                      "example": "z2@gmail.com"
                    },
                    "nickName": {
                      "type": "string",
                      "example": "aaa"
                    },
                    "profilePic": {
                      "type": "string",
                      "example": "上傳圖片回傳的URL"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2023-04-29T07:59:35.033Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2023-04-29T08:36:48.175Z"
                    },
                    "__v": {
                      "type": "number",
                      "example": 0
                    },
                    "birthday": {
                      "type": "string",
                      "example": "2023-04-29T08:20:13.000Z"
                    },
                    "phoneNumber": {
                      "type": "string",
                      "example": "0912345678"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "密碼強度",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "密碼長度至少8位、須包含數字與英文"
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "isOperational": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "402": {
            "description": "密碼不能為空",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "密碼不能為空"
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 402
                    },
                    "isOperational": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "406": {
            "description": "密碼不一致",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "密碼不一致"
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 406
                    },
                    "isOperational": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  },
...
}